The MerchantAnalyticsJob class in the Merchant Rating System has several key components:
MerchantPartitioner: This is a static inner class that extends the Partitioner class. It overrides the getPartition method to determine the partition based on the hash code of the key. This is used to distribute the data across different reducers. You can find this class in MerchantAnalyticsJob.java.

TransactionMapper: This is another static inner class that extends the Mapper class. It overrides the setup and map methods. The setup method is used to load the merchant ID and name into a cache, and the map method is used to map the input key/value pairs to output key/value pairs. You can also find this class in MerchantAnalyticsJob.java.

runMRJob: This method is used to set up and execute the MapReduce job. It configures the job, sets the input and output paths, sets the mapper, combiner, reducer, and partitioner classes, and then runs the job. This method is called in the run method of the Tool interface.

run: This method is an override of the run method in the Tool interface. It calls the runMRJob method to execute the MapReduce job.

LOGGER: This is a static final field used for logging. It's an instance of Logger from the LoggerFactory.

sdf: This is a static field used for formatting dates. It's an instance of SimpleDateFormat.

These components work together to process the merchant and transaction data using MapReduce.