The Transaction class in the file merchant-rating-system/src/main/java/org/paceWithMe/hadoop/helpers/Transaction.java implements the following methods:
getSegment(): This method returns the value of the segment field.
setSegment(String segment): This method sets the value of the segment field.
getTxId(): This method returns the value of the txId field.
setTxId(String txId): This method sets the value of the txId field.
getCustomerId(): This method returns the value of the customerId field.
setCustomerId(Long customerId): This method sets the value of the customerId field.
getMerchantId(): This method returns the value of the merchantId field.
setMerchantId(Long merchantId): This method sets the value of the merchantId field.
getTimeStamp(): This method returns the value of the timeStamp field.
setTimeStamp(String timeStamp): This method sets the value of the timeStamp field.
getInvoiceNumber(): This method returns the value of the invoiceNumber field.
setInvoiceNumber(String invoiceNumber): This method sets the value of the invoiceNumber field.
getInvoiceAmount(): This method returns the value of the invoiceAmount field.
setInvoiceAmount(float invoiceAmount): This method sets the value of the invoiceAmount field.